+++ system/src/dice/RollSD.mjs
@@ -58,7 +58,7 @@ export default class RollSD extends Roll {
 		}
 
 		// Roll damage for NPCs
-		if (data.actor?.type === "NPC") {
+		if (data.actor?.type === "NPC" && data.item.type === "NPC Attack") {
 			data = await this._rollNpcAttack(data);
 			if (!options.flavor) {
 				options.flavor = game.i18n.format(
@@ -86,16 +86,26 @@ export default class RollSD extends Roll {
 				}
 			}
 
-			// TODO This should really be set in the ItemSD rollSpell method...
 			// Spell? -> Set a target
 			if (data.item?.isSpell()) {
-				options.target = data.item.system.tier + 10;
+				// NPC spell
+				if (typeof data.item.system.dc !== "undefined") {
+					options.target = data.item.system.dc;
+					options.tier = data.item.system.dc - 10;
+					// system.tier needed for spell chat template
+					data.item.system.tier = options.tier;
+				}
+				// player spell
+				else {
+					options.target = data.item.system.tier + 10;
+					options.tier = data.item.system.tier;
+				}
 				if (!options.flavor) {
 					options.flavor = game.i18n.format(
 						"SHADOWDARK.chat.spell_roll.title",
 						{
 							name: data.item.name,
-							tier: data.item.system.tier,
+							tier: options.tier,
 							spellDC: options.target,
 						}
 					);
@@ -270,21 +280,42 @@ export default class RollSD extends Roll {
 	 * @returns {object}		- Returns the data object, with additional roll evaluations
 	 */
 	static async _rollNpcAttack(data) {
-		// Get the damage Die
-		let numDice = (data.item.system.damage.numDice)
-			? data.item.system.damage.numDice
-			: 1;
-		const damageDie = data.item.system.damage.value;
+		let baseDamageFormula = data.item.system.damage.value ?? "";
+		baseDamageFormula = baseDamageFormula.trim();
+
+		// Default to 1 damage if no damage formula has been set
+		baseDamageFormula = baseDamageFormula === "" ? "1" : baseDamageFormula;
 
 		// Get bonus damage
 		data.damageBonus = data.item.system.bonuses.damageBonus;
 		if (data.damageBonus) data.damageParts.push("@damageBonus");
 
-		if ( data.rolls.main.critical !== "failure" ) {
-			if ( data.rolls.main.critical === "success" ) numDice
-				*= parseInt(data.item.system.bonuses.critical.multiplier, 10);
+		if (data.rolls.main.critical !== "failure") {
+			// if (data.rolls.main.critical === "success") {
+			// We only support multiplication of the first damage dice,
+			// which is probably enough. None of the core game NPC attacks
+			// involve multiple dice parts
+			//
+			if (baseDamageFormula !== "0") {
+				const parts = /^(\d*)d(.*)/.exec(baseDamageFormula);
 
-			const primaryParts = [`${numDice}${damageDie}`, ...data.damageParts];
+				let numDice = "1";
+				let formulaSuffix = "";
+				if (parts) {
+					numDice = parts[1] !== "" ? parts[1] : "1";
+					formulaSuffix = parts[2] ? parts[2] : "";
+				}
+
+				numDice = parseInt(numDice, 10);
+				numDice *= parseInt(data.item.system.bonuses.critical.multiplier, 10);
+
+				baseDamageFormula = formulaSuffix !== ""
+					? `${numDice}d${formulaSuffix}`
+					: `${numDice}`;
+			}
+			// }
+
+			const primaryParts = [baseDamageFormula, ...data.damageParts];
 
 			data.rolls.primaryDamage = await this._roll(primaryParts, data);
 		}
@@ -516,11 +547,11 @@ export default class RollSD extends Roll {
 			user: game.user.id,
 			speaker: speaker,
 			flags: {
-				isRoll: true,
-				rolls: rolls,
+				"isRoll": true,
+				"rolls": rolls,
 				"core.canPopout": true,
-				hasTarget: target !== false,
-				critical: rolls.main.critical,
+				"hasTarget": target !== false,
+				"critical": rolls.main.critical,
 			},
 		};
 		if (target) chatData.flags.success = rolls.main.roll.total >= target;