+++ system/src/sheets/NpcSheetSD.mjs
@@ -54,9 +54,18 @@ export default class NpcSheetSD extends ActorSheetSD {
 
 	/** @inheritdoc */
 	activateListeners(html) {
-		html.find("[data-action='show-feature']").click(
-			event => this._onShowFeature(event)
+		html.find("[data-action='item-use-ability']").click(
+			event => this._onUseAbility(event)
 		);
+
+		html.find("[data-action='cast-npc-spell']").click(
+			event => this._onCastSpell(event)
+		);
+
+		html.find(".toggle-lost").click(
+			event => this._onToggleLost(event)
+		);
+
 		// Handle default listeners last so system listeners are triggered first
 		super.activateListeners(html);
 	}
@@ -78,6 +87,8 @@ export default class NpcSheetSD extends ActorSheetSD {
 
 	async _prepareItems(context) {
 		const attacks = [];
+		const specials = [];
+		const spells = [];
 		const features = [];
 
 		const effects = {
@@ -92,11 +103,20 @@ export default class NpcSheetSD extends ActorSheetSD {
 		};
 
 		for (const i of this._sortAllItems(context)) {
+			// Push Attacks
 			if (i.type === "NPC Attack") {
 				const display = await this.actor.buildNpcAttackDisplays(i._id);
 				attacks.push({itemId: i._id, display});
 			}
-			if (i.type === "NPC Feature") {
+
+			// Push Specials
+			else if (i.type === "NPC Special Attack") {
+				const display = await this.actor.buildNpcSpecialDisplays(i._id);
+				specials.push({itemId: i._id, display});
+			}
+
+			// Push Features
+			else if (i.type === "NPC Feature") {
 				const description = await TextEditor.enrichHTML(
 					jQuery(i.system.description).text(),
 					{
@@ -116,6 +136,19 @@ export default class NpcSheetSD extends ActorSheetSD {
 					display,
 				});
 			}
+
+			// Push Spells
+			else if (i.type === "NPC Spell") {
+				i.description = await TextEditor.enrichHTML(
+					jQuery(i.system.description).text(),
+					{
+						async: true,
+					}
+				);
+				spells.push(i);
+			}
+
+			// Push Effects
 			else if (i.type === "Effect") {
 				const category = i.system.category;
 				effects[category].items.push(i);
@@ -123,13 +156,23 @@ export default class NpcSheetSD extends ActorSheetSD {
 		}
 
 		context.attacks = attacks;
+		context.specials = specials;
+		context.spells = spells;
 		context.features = features;
 		context.effects = effects;
 	}
 
-	async _onShowFeature(event) {
+	async _onUseAbility(event) {
 		event.preventDefault();
 		const itemId = $(event.currentTarget).data("item-id");
 		this.actor.useAbility(itemId);
+	}
+
+	async _onCastSpell(event) {
+		event.preventDefault();
+
+		const itemId = $(event.currentTarget).data("item-id");
+
+		this.actor.castNPCSpell(itemId);
 	}
 }