+++ utils/javascript.mjs
@@ -1,11 +1,9 @@
 import { rollup } from "rollup";
 import eslint from "gulp-eslint-new";
 import gulp from "gulp";
-import livereload from "gulp-livereload";
 import gulpIf from "gulp-if";
 import mergeStream from "merge-stream";
 import nodeResolve from "@rollup/plugin-node-resolve";
-import { delay } from "../system/src/testing/testUtils.mjs";
 
 const SRC_LINT_PATHS = ["./system/shadowdark.mjs", "./system/src/"];
 
@@ -39,7 +37,7 @@ function lintJavascript() {
 
 		return gulp
 			.src(src)
-			.pipe(eslint({ fix: false }))
+			.pipe(eslint({ fix: true }))
 			.pipe(eslint.format())
 			.pipe(
 				gulpIf(
@@ -53,17 +51,9 @@ function lintJavascript() {
 }
 export const lint = lintJavascript;
 
-// Wait
-//
-export const wait = async () => {
-	await delay(1500);
-	livereload.reload();
-};
-
 // Watch for file changes and lint when they do
 //
 export async function watchJavascriptUpdates() {
-	livereload.listen();
-	gulp.watch(SRC_LINT_PATHS, gulp.parallel(lint, compile, wait));
+	gulp.watch(SRC_LINT_PATHS, gulp.parallel(lint, compile));
 }
 export const watchUpdates = watchJavascriptUpdates;