+++ RELEASE_NOTES.md
@@ -1,3 +1,33 @@
+## v.1.x.x
+
+### Bugfixes
+* [#573] Fixed an issue with item attack bonuses not being calculated if added manually via the roll dialog.
+* [#580] Learning a spell from a scroll doesn't detect that the player cancelled the roll
+* [#581] Spellcasting ability used for learning spells hardcoded, this should be dynamic to support custom/homebrew classes
+* [#582] Grick token is named "Monster"
+* [#583] Regex parser in MonsterImporterSD.mjs can't handle attacks that have '+' in the name
+* [#597] Unable to drag Actor to hotbar and have it open character sheet
+
+### Enhancements
+* [#220] Added ability to drag attacks, spells, abilities, potions, and light sources to the Hotbar.
+* [#184] Added special attack type to NPCs. The monster importer now adds special attacks.
+* [#328] Added spellcasting abilities and spells to NPCs. The monster importer now adds spells.
+* [#566] Warn players when they're attempting to learn a spell that is from a different class
+* [#578] Standardize spellcasting iconography
+* [#585] Add predefined Talent effect that allows the addition of an extra damage die to weapons with specified tags (`weaponDamageExtraDieByProperty`)
+* [#586] Add predefined Talent effect that allows extra damage die for weapons with specified tags to be increased by category (`weaponDamageExtraDieImprovementByProperty`)
+* [#587] Add predefined Talent effect that allows damage die for weapons with specified tags to be increased by category (`weaponDamageDieImprovementByProperty`)
+* [#596] Switch NPC number of attacks to text in order to support roll macros within attack strings
+
+*Please welcome Prototype to the Shadowdark system development team*
+
+*Many thanks to chrpow for contributing code included in this build*
+
+### Chores
+* Improvements to GitHub Workflows around building a release
+
+*Many thanks to chrpow for contributing code included in this build*
+
 ## v.1.5.1
 
 ### Bugfixes